@@ -1,5 +1,6 @@
-import { GLOBAL_ERROR, GET_ORDERS, GET_ORDER, MULTI_ORDER_LOADING, SINGLE_ORDER_LOADING } from "../types";
+import { GLOBAL_ERROR, GET_ORDERS, GET_ORDER, MULTI_ORDER_LOADING, SINGLE_ORDER_LOADING, TOGGLE_ORDER_APPROVAL, SUCCESS, TOGGLE_TOBERETURN_ORDER, CANCEL_ORDER } from "../types";
 import api from "../../utils/api";
+import axios from 'axios'
 
 
 export const getOrders = (id,page,perPage,status='', keyword='') => async (dispatch) => {
@@ -26,6 +27,58 @@ export const getOrder = (id,order_id) => async (dispatch) => {
     });
   } catch (err) {
     console.log("****order_actions/getOrder****", err);
+    dispatch({ type: GLOBAL_ERROR, payload: err || "Not Found" });
+  }
+};
+
+export const toggleOrderApproval = (id, order_id) => async (dispatch) => {
+    try {
+    const res = await api.patch(`/order/toggle-order-approval/${id}/${order_id}`);
+    dispatch({
+      type: TOGGLE_ORDER_APPROVAL,
+      payload: res.data,
+    });
+    dispatch({
+      type: SUCCESS,
+      payload: `Order status has sucessfully changed to ${res.data.status.currentStatus}`,
+    });
+  } catch (err) {
+    console.log("****order_actions/toggleOrderApproval****", err);
+    dispatch({ type: GLOBAL_ERROR, payload: err || "Not Found" });
+  }
+};
+
+export const toggletobeReturnOrder = (id, order_id, remark = '', returnedAmount = '' ) => async (dispatch) => {
+  const body = JSON.stringify({ remark, returnedAmount });//required only if complete to tobereturned
+  try {
+    const res = await api.patch(`/order/toggle-order-to-get-return/${id}/${order_id}`,body);
+    dispatch({
+      type: TOGGLE_TOBERETURN_ORDER,
+      payload: res.data,
+    });
+    dispatch({
+      type: SUCCESS,
+      payload: `Order status has sucessfully changed to ${res.data.order.status.currentStatus}`,
+    });
+  } catch (err) {
+    console.log("****order_actions/toggletobeReturnOrder****", err);
+    dispatch({ type: GLOBAL_ERROR, payload: err || "Not Found" });
+  }
+};
+
+export const cancelOrder = (id, order_id) => async (dispatch) => {
+  try {
+    const res = await api.patch(`/order/cancel-order/${id}/${order_id}`);
+    dispatch({
+      type: CANCEL_ORDER,
+      payload: res.data,
+    });
+    dispatch({
+      type: SUCCESS,
+      payload: 'Order cancelled sucessfully',
+    });
+  } catch (err) {
+    console.log("****order_actions/orderCancel****", err);
     dispatch({ type: GLOBAL_ERROR, payload: err || "Not Found" });
   }
 };
